requirements.txt
fastapi
uvicorn
pydantic
PyMuPDF
google-generativeai
python-dotenv
requests
sqlalchemy
psycopg2-binary
